pyproject.toml
[tool.poetry]
name = "aws-transcribe-bedrock"
version = "0.1.0"
description = "Voice Assistant with AWS Transcribe and Bedrock - A web-based interface for voice input transcription and text optimization"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false
packages = [{include = "voice_assistant", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.69"
gradio = "^5.22.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true